+++ src/zope/mimetype/tests/__init__.py
@@ -23,7 +23,7 @@ class ISampleContentTypeOne(zope.interface.Interface):
     """This is a sample content type interface."""
 
 
-ISampleContentTypeOne.setTaggedValue("title", u"Type One")
+ISampleContentTypeOne.setTaggedValue("title", "Type One")
 ISampleContentTypeOne.setTaggedValue("extensions", [])
 ISampleContentTypeOne.setTaggedValue("mimeTypes", ["type/one", "type/foo"])
 
@@ -36,7 +36,7 @@ class ISampleContentTypeTwo(zope.interface.Interface):
     """This is a sample content type interface."""
 
 
-ISampleContentTypeTwo.setTaggedValue("title", u"Type Two")
+ISampleContentTypeTwo.setTaggedValue("title", "Type Two")
 ISampleContentTypeTwo.setTaggedValue("mimeTypes", [".ct2", ".ct3"])
 ISampleContentTypeTwo.setTaggedValue("mimeTypes", ["type/two"])