+++ src/zope/mimetype/tests/test_codec.py
@@ -15,9 +15,6 @@
 Tests for codec.py.
 
 """
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
 
 import unittest