+++ src/zope/mimetype/tests/test_doctests.py
@@ -12,31 +12,11 @@
 ##############################################################################
 """Test harness for `zope.mimetype` doctests..
 """
-from __future__ import print_function
 
 import doctest
-import re
 import unittest
 
 from zope.component import testing
-from zope.testing import renormalizing
-
-
-checker = renormalizing.RENormalizing([
-    # Python 3 unicode removed the "u".
-    (re.compile("u('.*?')"),
-     r"\1"),
-    (re.compile('u(".*?")'),
-     r"\1"),
-    # Python 3 bytes adds the "b".
-    (re.compile("b('.*?')"),
-     r"\1"),
-    (re.compile('b(".*?")'),
-     r"\1"),
-    # Python 3 renames builtins.
-    (re.compile("__builtin__"),
-     r"builtins"),
-])
 
 
 def optional_test_suite_setUp(required_module):
@@ -57,44 +37,33 @@ def test_suite():
     return unittest.TestSuite((
         doctest.DocFileSuite(
             '../retrieving_mime_types.rst',
-            setUp=testing.setUp, tearDown=testing.tearDown,
-            checker=checker),
+            setUp=testing.setUp, tearDown=testing.tearDown),
         doctest.DocFileSuite(
             '../event.rst',
-            setUp=testing.setUp, tearDown=testing.tearDown,
-            globs={'print_function': print_function},
-            checker=checker),
+            setUp=testing.setUp, tearDown=testing.tearDown),
         doctest.DocFileSuite(
             '../source.rst',
             setUp=optional_test_suite_setUp('zope.browser'),
-            tearDown=testing.tearDown,
-            checker=checker),
+            tearDown=testing.tearDown),
         doctest.DocFileSuite(
-            '../constraints.rst',
-            checker=checker),
+            '../constraints.rst'),
         doctest.DocFileSuite(
             '../contentinfo.rst',
-            setUp=testing.setUp, tearDown=testing.tearDown,
-            checker=checker),
+            setUp=testing.setUp, tearDown=testing.tearDown),
         doctest.DocFileSuite(
-            '../typegetter.rst',
-            checker=checker),
+            '../typegetter.rst'),
         doctest.DocFileSuite(
-            '../utils.rst',
-            checker=checker),
+            '../utils.rst'),
         doctest.DocFileSuite(
             '../widget.rst',
             setUp=optional_test_suite_setUp('zope.publisher'),
-            tearDown=testing.tearDown,
-            checker=checker),
+            tearDown=testing.tearDown),
         doctest.DocFileSuite(
             '../codec.rst',
             optionflags=doctest.NORMALIZE_WHITESPACE | doctest.ELLIPSIS,
-            checker=checker,
         ),
         doctest.DocFileSuite(
             '../configure.rst',
             setUp=testing.setUp, tearDown=testing.tearDown,
-            checker=checker
         ),
     ))